Reorg save_* related efuns.
